Add type definitions for custom events
