+++ .github/workflows/tests.yml
@@ -24,6 +24,9 @@ jobs:
     - name: Lint CSS
       run: npm run lint:css
 
+    - name: Lint typings
+      run: npm run lint:types
+
   unit-tests:
     runs-on: ubuntu-latest
