+++ package.json
@@ -29,9 +29,10 @@
     "check-version": "magi check-version",
     "debug": "web-test-runner test/*.test.js --watch",
     "dist": "rimraf dist && npm run analyze && rollup -c rollup.config.js && cp analysis.json dist",
-    "lint": "npm run lint:js && npm run lint:css",
+    "lint": "npm run lint:js && npm run lint:css && npm run lint:types",
     "lint:css": "stylelint src/*.js theme/**/*-styles.js",
     "lint:js": "eslint src theme test",
+    "lint:types": "tsc",
     "prestart": "npm run analyze",
     "preversion": "magi update-version",
     "screenshots": "hermione test/visual/test.js --update-refs",
@@ -84,6 +85,7 @@
     "sinon": "^9.2.1",
     "stylelint": "^13.8.0",
     "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-vaadin": "^0.2.10"
+    "stylelint-config-vaadin": "^0.2.10",
+    "typescript": "^4.1.3"
   }
 }
