+++ vaadin-notification.d.ts
@@ -1,1 +1,2 @@
 export * from './src/vaadin-notification.js';
+export * from './src/interfaces';
